+++ src/handlers/GSMConnectionHandler.cpp
@@ -19,10 +19,10 @@
    INCLUDE
  ******************************************************************************/
 
-#include "ConnectionHandlerDefinitions.h"
+#include "definitions/ConnectionHandlerDefinitions.h"
 
 #ifdef BOARD_HAS_GSM /* Only compile if this is a board with GSM */
-#include "GSMConnectionHandler.h"
+#include "handlers/GSMConnectionHandler.h"
 
 /******************************************************************************
    CONSTANTS
