+++ src/handlers/GSMConnectionHandler.h
@@ -15,57 +15,37 @@
    a commercial license, send an email to license@arduino.cc.
 */
 
-#ifndef ARDUINO_WIFI_CONNECTION_HANDLER_H_
-#define ARDUINO_WIFI_CONNECTION_HANDLER_H_
+#ifndef GSM_CONNECTION_MANAGER_H_
+#define GSM_CONNECTION_MANAGER_H_
 
 /******************************************************************************
    INCLUDE
  ******************************************************************************/
 
-#include "ConnectionHandlerInterface.h"
+#include "interface/ConnectionHandlerInterface.h"
 
-#ifdef ARDUINO_SAMD_MKR1000
-  #include <WiFi101.h>
-  #include <WiFiUdp.h>
-#elif defined(ARDUINO_SAMD_MKRWIFI1010) || defined(ARDUINO_SAMD_NANO_33_IOT) || \
-  defined(ARDUINO_AVR_UNO_WIFI_REV2) || defined (ARDUINO_NANO_RP2040_CONNECT)
-  #include <WiFiNINA.h>
-  #include <WiFiUdp.h>
-#elif defined(ARDUINO_PORTENTA_H7_M7) || defined(ARDUINO_PORTENTA_H7_M7) || \
-  defined(ARDUINO_NICLA_VISION) || defined(ARDUINO_OPTA) || defined(ARDUINO_GIGA)
-  #include <WiFi.h>
-  #include <WiFiUdp.h>
-#elif defined(ARDUINO_PORTENTA_C33)
-  #include <WiFiC3.h>
-  #include <WiFiUdp.h>
-#elif defined(ARDUINO_ARCH_ESP8266)
-  #include <ESP8266WiFi.h>
-  #include <WiFiUdp.h>
-#elif defined(ARDUINO_ARCH_ESP32)
-  #include <WiFi.h>
-  #include <WiFiUdp.h>
-#elif defined(ARDUINO_UNOR4_WIFI)
-  #include <WiFiS3.h>
+#if defined(ARDUINO_SAMD_MKRGSM1400)
+  #include <MKRGSM.h>
 #endif
 
-#ifndef BOARD_HAS_WIFI
-  #error "Board doesn't support WIFI"
+#ifndef BOARD_HAS_GSM
+  #error "Board doesn't support GSM"
 #endif
 
 /******************************************************************************
    CLASS DECLARATION
  ******************************************************************************/
 
-class WiFiConnectionHandler : public ConnectionHandler
+class GSMConnectionHandler : public ConnectionHandler
 {
   public:
 
-    WiFiConnectionHandler(char const * ssid, char const * pass, bool const keep_alive = true);
+    GSMConnectionHandler(const char * pin, const char * apn, const char * login, const char * pass, bool const keep_alive = true);
 
 
     virtual unsigned long getTime() override;
-    virtual Client & getClient() override { return _wifi_client; }
-    virtual UDP & getUDP() override { return _wifi_udp; }
+    virtual Client & getClient() override { return _gsm_client; };
+    virtual UDP & getUDP() override { return _gsm_udp; };
 
 
   protected:
@@ -76,13 +56,18 @@ class WiFiConnectionHandler : public ConnectionHandler
     virtual NetworkConnectionState update_handleDisconnecting() override;
     virtual NetworkConnectionState update_handleDisconnected () override;
 
+
   private:
 
-    char const * _ssid;
-    char const * _pass;
+    const char * _pin;
+    const char * _apn;
+    const char * _login;
+    const char * _pass;
 
-    WiFiUDP _wifi_udp;
-    WiFiClient _wifi_client;
+    GSM _gsm;
+    GPRS _gprs;
+    GSMUDP _gsm_udp;
+    GSMClient _gsm_client;
 };
 
-#endif /* ARDUINO_WIFI_CONNECTION_HANDLER_H_ */
+#endif /* #ifndef GSM_CONNECTION_MANAGER_H_ */
